backend/requirements.txt
fastapi
uvicorn
numpy==1.26.4
openai
python-dotenv
faiss-cpu==1.7.4
setuptools>=61.0.0